server/requirements.txt
Flask>=3.0
flask-pymongo>=2.4
Flask-JWT-Extended>=4.6
Flask-Cors>=4.0
APScheduler>=3.10
python-dotenv>=1.0
marshmallow>=3.21
passlib[bcrypt]>=1.7
Flask-APScheduler>=1.13 